+++ docs/about/whats_new.rst
@@ -7,6 +7,9 @@ What's New
 
 v1.9.next
 =========
+- Improve error message for mismatch between dataset metadata and product signature (:pull:`1472`)
+- Mark ``--confirm-ignore-lineage``, ``--auto-add-lineage``, and ``--verify-lineage`` as deprecated or to be deprecated (:pull:`1472`)
+- Default delta values in ``archive_less_mature`` and ``find_less_mature`` (:pull:`1472`)
 
 - External Lineage API (:pull:`#1401`)
 - Add lineage support to index clone operation (:pull:`#1429`)